--- a/core/trino-main/src/test/java/io/trino/execution/buffer/BenchmarkBlockSerde.java
+++ b/core/trino-main/src/test/java/io/trino/execution/buffer/BenchmarkBlockSerde.java
@@ -27,6 +27,7 @@ import io.trino.spi.block.TestingBlockEncodingSerde;
 import io.trino.spi.type.DecimalType;
 import io.trino.spi.type.Decimals;
 import io.trino.spi.type.Int128;
+import io.trino.spi.type.RowType;
 import io.trino.spi.type.SqlDecimal;
 import io.trino.spi.type.Type;
 import io.trino.spi.type.VarcharType;
@@ -42,7 +43,9 @@ import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 import org.openjdk.jmh.annotations.Warmup;
 import org.testng.annotations.Test;
+import oshi.util.tuples.Pair;
 
+import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Random;
@@ -176,6 +179,18 @@ public class BenchmarkBlockSerde
         return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
     }
 
+    @Benchmark
+    public Object serializeRow(RowTypeBenchmarkData data)
+    {
+        return serializePages(data);
+    }
+
+    @Benchmark
+    public Object deserializeRow(RowTypeBenchmarkData data)
+    {
+        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+    }
+
     private static List<Slice> serializePages(BenchmarkData data)
     {
         PagesSerdeContext context = new PagesSerdeContext();
@@ -199,35 +214,7 @@ public class BenchmarkBlockSerde
 
             Iterator<?> values = createValues(ROWS, valueGenerator, nullChance);
             while (values.hasNext()) {
-                Object value = values.next();
-                if (value == null) {
-                    blockBuilder.appendNull();
-                }
-                else if (BIGINT.equals(type)) {
-                    BIGINT.writeLong(blockBuilder, ((Number) value).longValue());
-                }
-                else if (Decimals.isLongDecimal(type)) {
-                    type.writeObject(blockBuilder, Int128.valueOf(((SqlDecimal) value).toBigDecimal().unscaledValue()));
-                }
-                else if (type instanceof VarcharType) {
-                    Slice slice = truncateToLength(utf8Slice((String) value), type);
-                    type.writeSlice(blockBuilder, slice);
-                }
-                else if (TIMESTAMP_PICOS.equals(type)) {
-                    TIMESTAMP_PICOS.writeObject(blockBuilder, value);
-                }
-                else if (INTEGER.equals(type)) {
-                    blockBuilder.writeInt((int) value);
-                }
-                else if (SMALLINT.equals(type)) {
-                    blockBuilder.writeShort((short) value);
-                }
-                else if (TINYINT.equals(type)) {
-                    blockBuilder.writeByte((byte) value);
-                }
-                else {
-                    throw new IllegalArgumentException("Unsupported type " + type);
-                }
+                writeValue(type, values.next(), blockBuilder);
                 pageBuilder.declarePosition();
                 if (pageBuilder.isFull()) {
                     pagesBuilder.add(pageBuilder.build());
@@ -244,6 +231,51 @@ public class BenchmarkBlockSerde
             writePages(pagesSerde, new OutputStreamSliceOutput(sliceOutput), pages.iterator());
 
             setup(sliceOutput.slice(), pagesSerde, pages);
+        }
+
+        private void writeValue(Type type, Object value, BlockBuilder blockBuilder)
+        {
+            if (value == null) {
+                blockBuilder.appendNull();
+            }
+            else if (BIGINT.equals(type)) {
+                BIGINT.writeLong(blockBuilder, ((Number) value).longValue());
+            }
+            else if (Decimals.isLongDecimal(type)) {
+                type.writeObject(blockBuilder, Int128.valueOf(((SqlDecimal) value).toBigDecimal().unscaledValue()));
+            }
+            else if (type instanceof VarcharType) {
+                Slice slice = truncateToLength(utf8Slice((String) value), type);
+                type.writeSlice(blockBuilder, slice);
+            }
+            else if (TIMESTAMP_PICOS.equals(type)) {
+                TIMESTAMP_PICOS.writeObject(blockBuilder, value);
+            }
+            else if (INTEGER.equals(type)) {
+                blockBuilder.writeInt((int) value);
+            }
+            else if (SMALLINT.equals(type)) {
+                blockBuilder.writeShort((short) value);
+            }
+            else if (TINYINT.equals(type)) {
+                blockBuilder.writeByte((byte) value);
+            }
+            else if (type instanceof RowType) {
+                BlockBuilder row = blockBuilder.beginBlockEntry();
+                List<?> values = (List<?>) value;
+                if (values.size() != type.getTypeParameters().size()) {
+                    throw new IllegalArgumentException("Size of types and values must have the same size");
+                }
+                List<Pair<Type, Object>> pairs = new ArrayList<>();
+                for (int i = 0; i < type.getTypeParameters().size(); i++) {
+                    pairs.add(new Pair<Type, Object>(type.getTypeParameters().get(i), ((List<?>) value).get(i)));
+                }
+                pairs.forEach(p -> writeValue(p.getA(), p.getB(), row));
+                blockBuilder.closeEntry();
+            }
+            else {
+                throw new IllegalArgumentException("Unsupported type " + type);
+            }
         }
     }
 
@@ -366,6 +398,18 @@ public class BenchmarkBlockSerde
             DynamicSliceOutput sliceOutput = new DynamicSliceOutput(0);
             writePages(pagesSerde, new OutputStreamSliceOutput(sliceOutput), pages.listIterator());
             setup(sliceOutput.slice(), pagesSerde, pages);
+        }
+    }
+
+    @State(Thread)
+    public static class RowTypeBenchmarkData
+            extends TypeBenchmarkData
+    {
+        @Setup
+        public void setup()
+        {
+            RowType type = RowType.anonymous(ImmutableList.of(BIGINT));
+            super.setup(type, (random -> BenchmarkDataGenerator.randomRow(type.getTypeParameters(), random)));
         }
     }