--- a/core/trino-spi/src/main/java/io/trino/spi/block/RowBlock.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/block/RowBlock.java
@@ -46,15 +46,10 @@ public class RowBlock
     public static Block fromFieldBlocks(int positionCount, Optional<boolean[]> rowIsNullOptional, Block[] fieldBlocks)
     {
         boolean[] rowIsNull = rowIsNullOptional.orElse(null);
-        int[] fieldBlockOffsets = new int[positionCount + 1];
-        if (rowIsNull == null) {
-            // Fast-path create identity field block offsets from position only
-            for (int position = 0; position < fieldBlockOffsets.length; position++) {
-                fieldBlockOffsets[position] = position;
-            }
-        }
-        else {
+        int[] fieldBlockOffsets = null;
+        if (rowIsNull != null) {
             // Check for nulls when computing field block offsets
+            fieldBlockOffsets = new int[positionCount + 1];
             fieldBlockOffsets[0] = 0;
             for (int position = 0; position < positionCount; position++) {
                 fieldBlockOffsets[position + 1] = fieldBlockOffsets[position] + (rowIsNull[position] ? 0 : 1);
@@ -63,6 +58,7 @@ public class RowBlock
             if (fieldBlockOffsets[positionCount] == positionCount) {
                 // No nulls encountered, discard the null mask
                 rowIsNull = null;
+                fieldBlockOffsets = null;
             }
         }
 
@@ -73,13 +69,13 @@ public class RowBlock
     /**
      * Create a row block directly without per element validations.
      */
-    static RowBlock createRowBlockInternal(int startOffset, int positionCount, @Nullable boolean[] rowIsNull, int[] fieldBlockOffsets, Block[] fieldBlocks)
+    static RowBlock createRowBlockInternal(int startOffset, int positionCount, @Nullable boolean[] rowIsNull, @Nullable int[] fieldBlockOffsets, Block[] fieldBlocks)
     {
         validateConstructorArguments(startOffset, positionCount, rowIsNull, fieldBlockOffsets, fieldBlocks);
         return new RowBlock(startOffset, positionCount, rowIsNull, fieldBlockOffsets, fieldBlocks);
     }
 
-    private static void validateConstructorArguments(int startOffset, int positionCount, @Nullable boolean[] rowIsNull, int[] fieldBlockOffsets, Block[] fieldBlocks)
+    private static void validateConstructorArguments(int startOffset, int positionCount, @Nullable boolean[] rowIsNull, @Nullable int[] fieldBlockOffsets, Block[] fieldBlocks)
     {
         if (startOffset < 0) {
             throw new IllegalArgumentException("arrayOffset is negative");
@@ -93,8 +89,11 @@ public class RowBlock
             throw new IllegalArgumentException("rowIsNull length is less than positionCount");
         }
 
-        requireNonNull(fieldBlockOffsets, "fieldBlockOffsets is null");
-        if (fieldBlockOffsets.length - startOffset < positionCount + 1) {
+        if ((rowIsNull == null) != (fieldBlockOffsets == null)) {
+            throw new IllegalArgumentException("When rowIsNull is (non) null then fieldBlockOffsets should be (non) null as well");
+        }
+
+        if (fieldBlockOffsets != null && fieldBlockOffsets.length - startOffset < positionCount + 1) {
             throw new IllegalArgumentException("fieldBlockOffsets length is less than positionCount");
         }
 
@@ -116,7 +115,7 @@ public class RowBlock
      * Use createRowBlockInternal or fromFieldBlocks instead of this method.  The caller of this method is assumed to have
      * validated the arguments with validateConstructorArguments.
      */
-    private RowBlock(int startOffset, int positionCount, @Nullable boolean[] rowIsNull, int[] fieldBlockOffsets, Block[] fieldBlocks)
+    private RowBlock(int startOffset, int positionCount, @Nullable boolean[] rowIsNull, @Nullable int[] fieldBlockOffsets, Block[] fieldBlocks)
     {
         super(fieldBlocks.length);
 
@@ -136,6 +135,7 @@ public class RowBlock
     }
 
     @Override
+    @Nullable
     protected int[] getFieldBlockOffsets()
     {
         return fieldBlockOffsets;
@@ -176,8 +176,8 @@ public class RowBlock
         long sizeInBytes = getBaseSizeInBytes();
         boolean hasUnloadedBlocks = false;
 
-        int startFieldBlockOffset = fieldBlockOffsets[startOffset];
-        int endFieldBlockOffset = fieldBlockOffsets[startOffset + positionCount];
+        int startFieldBlockOffset = fieldBlockOffsets != null ? fieldBlockOffsets[startOffset] : startOffset;
+        int endFieldBlockOffset = fieldBlockOffsets != null ? fieldBlockOffsets[startOffset + positionCount] : startOffset + positionCount;
         int fieldBlockLength = endFieldBlockOffset - startFieldBlockOffset;
 
         for (Block fieldBlock : fieldBlocks) {