@sopel39 2022-01-31
mind automation (maven checks)